ytdl/requirements.txt
torch
torchaudio
pyannote.audio
faster-whisper
